requirements.txt
asgiref==3.3.1
astroid==2.5
boto3==1.17.43
botocore==1.20.43
dj-database-url==0.5.0
Django==3.1.7
django-storages==1.11.1
djangorestframework==3.12.2
gunicorn==20.1.0
isort==5.7.0
jmespath==0.10.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
Pillow==8.1.2
psycopg2-binary==2.8.6
pylint==2.7.1
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
s3transfer==0.3.6
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
whitenoise==5.2.0
wrapt==1.12.1
